introduce WorkspaceContentService to handle workspace file loading
