+++ vscode/microsoft-kiota/src/modules/workspace/index.ts
@@ -1,0 +1,9 @@
+import { ClientOrPluginProperties } from "../../kiotaInterop";
+import WorkspaceContentService from "./workspaceContentService";
+
+export interface WorkspaceContent {
+  version: string;
+  clients: Record<string, ClientOrPluginProperties>;
+  plugins: Record<string, ClientOrPluginProperties>;
+}
+export { WorkspaceContentService };
